Fix ETXTBSY in copyAssets
